Move latest code into lib.
