+++ lib/setup.cfg
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 [metadata]
-name = multiverse
-version = 0.0.1
+name = unicorn-multiverse
+version = 0.0.2
 author = Philip Howard
 author_email = phil@gadgetoid.com
 description = Multi USB serial display driver
